Refactor benchmarks infra, update docs and results
- Overhaul Update-BenchmarkHistory.ps1: add dedup/version logic, culture-invariant parsing, -Force, and improved messaging
- BenchmarkHistory.md: new format, new results, only user [Params] as Options, remove infra columns
- Benchmarks now use temp working dirs per iteration for clean instrumentation; add [IterationSetup]/[GlobalCleanup]
- Remove redundant benchmarks; clarify comments and structure
- Program.cs: enforce invariant culture, semicolon CSV, remove deleted benchmarks
- Update documentation for new workflow and implementation plan
- General cleanup for reliability, reproducibility, and clarity
